Preserve files for hits on any intel types.
So far only intel hits on files that have already been extracted (i.e.
on FILE_HASH) triggerd preservation. Now, all hits on a file or a
connection involved in the file transfer will trigger preservation of
the file. Using Intel::preserve_filter the types for which to preserve
files can be selected explicitly.